[reddit] Added cake day detection to .redditor command
Added a 'Cake day' section, the text of which is pink after the redditor's
name if it is their cake day.

The definition of a user's cake day is taken from this post by a reddit
admin: http://redd.it/1v6xam.  The cake day is defined as the 24 hours
following the exact time, day and month of the user account's creation.  If the
account was created on February 29th and it is not a leap year, then the cake
day is taken to be on March 1st at the same time.

On reddit.com, the cake day can be 'held' for a number of days if the user does
not log in.  The log in habits of users are not exposed by the reddit API, so
this holding of cake days is not detected.