[reddit] Added .setsafeforwork command
The .setsafeforwork command takes an argument of either 'true'
or 'false' and sets the value of the sfw entry in the
specified channel's channel_values to the correspomding boolean.

The sfw entry is intended to allow handling of NSFW content, primarily
from reddit. As of this commit, the sfw value is not used at all in the
default reddit module.